page tree node fixes --- in progress
